feat: dynamic cart overview
